Functionality for Latencies and Accuracies
Added functionality for calculating latencies and accuracies for each character in a given response which are entered as lists in the Results table linked to the position of the response in the test. Right now it has a bug that causes it to not calculate the latency for the first character inputted, but that will be fixed.